feat: add ServerSessionRegistry so sessions outlive a sweep
