+++ AppShared/Sources/AppShared/Networking/SyncEngine.swift
@@ -30,9 +30,7 @@ import os
 @MainActor
 @Observable
 public final class SyncEngine {
-  @ObservationIgnored private let database: Database
   @ObservationIgnored private let manager: ConnectionManager
-  @ObservationIgnored private let mode: ApiRepository.Mode
   /// Live raw path cost read for the observation-driven initial-sync path (the
   /// lifecycle-triggered sweep receives it as a parameter instead). Raw, not a
   /// decision — each server combines it with its own `syncOverCellular` via
@@ -40,70 +38,39 @@ public final class SyncEngine {
   @ObservationIgnored private let pathCost:
     @MainActor () -> (isExpensive: Bool, isConstrained: Bool)
 
-  /// One session per server, retained across sweeps.
-  ///
-  /// This replaces the parallel `inFlight` / `lastSweep` dictionaries that used
-  /// to live here: both were per-server state keyed by UUID, which is precisely
-  /// what a session *is*. Retaining them also means a server's repository —
-  /// and therefore its `activeFills` dedupe — survives from one sweep to the
-  /// next instead of being rebuilt and forgotten each time.
-  ///
-  /// The active server is never keyed here: it is `DocumentStore`'s to drive.
-  @ObservationIgnored private var sessions: [UUID: ServerSession] = [:]
+  /// Where the per-server sessions live. The engine borrows them; it does not
+  /// own them, and it no longer keeps any per-server state of its own.
+  @ObservationIgnored private let registry: ServerSessionRegistry
   /// Whole-sweep single-flight, coalescing concurrent lifecycle triggers.
   @ObservationIgnored private var sweepTask: Task<Void, Never>?
-  /// Server IDs seen at the last observation tick, to detect newly-added ones.
-  @ObservationIgnored private var knownServerIDs: Set<UUID> = []
-  @ObservationIgnored private var observationTask: Task<Void, Never>?
 
   /// Background warmth cadence for inactive servers — deliberately looser than
   /// the active path's 300 s reconcile throttle; these servers aren't on screen.
   @ObservationIgnored private let inactiveThrottle: TimeInterval = 900
 
   public init(
-    database: Database,
+    registry: ServerSessionRegistry,
     manager: ConnectionManager,
-    pathCost: @escaping @MainActor () -> (isExpensive: Bool, isConstrained: Bool),
-    mode: ApiRepository.Mode = Bundle.main.appConfiguration.mode
+    pathCost: @escaping @MainActor () -> (isExpensive: Bool, isConstrained: Bool)
   ) {
-    self.database = database
+    self.registry = registry
     self.manager = manager
     self.pathCost = pathCost
-    self.mode = mode
-  }
-
-  deinit {
-    observationTask?.cancel()
   }
 
   // MARK: - Public API
 
-  /// Begin observing the `server` table (via `manager.connections`, which is the
-  /// sole projection of it). A newly-appeared, non-active server triggers a
-  /// throttle-exempt initial sync.
+  /// Start reacting to the server table: a newly-appeared, non-active server
+  /// gets a throttle-exempt initial sync.
   ///
-  /// This is wired to the *observation*, not the "Add server" UI action, so a
-  /// Stage-12 UBKVS `server` upsert flows into the initial-sync path for free.
-  /// Removed servers need no cache work — the `server` row delete FK-cascades
-  /// the whole cache; the engine just drops their session.
+  /// The observation itself belongs to ``ServerSessionRegistry`` — it owns
+  /// session lifetime, and one observer of that table is the whole point. The
+  /// engine only supplies the *policy* for what an addition deserves.
   public func start() {
-    guard observationTask == nil else { return }
-    // Seed with the current set so the observation only reacts to *new* servers;
-    // the servers present at launch are handled by the explicit sweep below.
-    knownServerIDs = Set(manager.connections.keys)
-    observationTask = Task { @MainActor [weak self] in
-      while !Task.isCancelled {
-        let (stream, continuation) = AsyncStream<Void>.makeStream()
-        withObservationTracking {
-          _ = self?.manager.connections.keys
-        } onChange: {
-          continuation.yield()
-          continuation.finish()
-        }
-        self?.handleConnectionsChanged()
-        for await _ in stream { break }
-      }
+    registry.onServersChanged = { [weak self] current, previous in
+      self?.handleServersChanged(current: current, previous: previous)
     }
+    registry.start()
   }
 
   /// Sweep every inactive server once, sequentially (active-first is implicit —
@@ -140,7 +107,7 @@ public final class SyncEngine {
     let actions = SyncPlan.inactiveActions(
       connections: snapshots,
       activeID: manager.activeConnectionId,
-      lastSweep: userInitiated ? [:] : lastSuccessfulSyncs(),
+      lastSweep: userInitiated ? [:] : registry.lastSuccessfulSyncs(),
       now: Date(),
       throttle: inactiveThrottle,
       isExpensive: isExpensive,
@@ -160,15 +127,12 @@ public final class SyncEngine {
     Logger.sync.info("Inactive sweep complete")
   }
 
-  private func handleConnectionsChanged() {
-    let current = Set(manager.connections.keys)
-    // Drop the session of any server whose row vanished (cache already cascaded).
-    for id in knownServerIDs.subtracting(current) {
-      sessions.removeValue(forKey: id)?.invalidate()
-    }
+  /// Policy for a server that just appeared. Session creation and teardown
+  /// already happened in the registry; all that is left is deciding what the
+  /// addition deserves, which is `SyncPlan`'s call.
+  private func handleServersChanged(current: Set<UUID>, previous: Set<UUID>) {
     let added = SyncPlan.newlyAdded(
-      current: current, known: knownServerIDs, activeID: manager.activeConnectionId)
-    knownServerIDs = current
+      current: current, known: previous, activeID: manager.activeConnectionId)
     if !added.isEmpty {
       Logger.sync.info("Observed \(added.count) new server(s); kicking initial sync")
     }
@@ -190,31 +154,12 @@ public final class SyncEngine {
   // MARK: - Per-server execution
 
   private func runAction(_ action: SyncServerAction, stored: StoredConnection) async {
-    let session = session(for: stored.id)
+    let session = registry.session(for: stored.id)
     if action.needsAuthOnly {
       session.markNeedsAuth(stored)
       return
     }
     await session.sync(stored: stored, runHeavyFill: action.runHeavyFill)
-  }
-
-  // MARK: - Sessions
-
-  /// The server's session, created on first use. Sessions are cheap — a
-  /// repository is only assembled when something actually syncs.
-  private func session(for id: UUID) -> ServerSession {
-    if let existing = sessions[id] {
-      return existing
-    }
-    let session = ServerSession(
-      serverID: id, database: database, manager: manager, mode: mode)
-    sessions[id] = session
-    return session
-  }
-
-  /// The throttle input `SyncPlan` consumes, projected out of the sessions.
-  private func lastSuccessfulSyncs() -> [UUID: Date] {
-    sessions.compactMapValues(\.lastSuccessfulSync)
   }
 
   // MARK: - Helpers
